producer/requirements.txt
pandas
celery
psycopg2-binary
langchain_community
boto3